+++ lib/whois/record/parser/whois.fi.rb
@@ -1,0 +1,76 @@
+#--
+# Ruby Whois
+#
+# An intelligent pure Ruby WHOIS client and parser.
+#
+# Copyright (c) 2009-2012 Simone Carletti <weppos@weppos.net>
+#++
+
+
+require 'whois/record/parser/base'
+
+
+module Whois
+  class Record
+    class Parser
+
+      # Parser for the whois.fi server.
+      #
+      # @note This parser is just a stub and provides only a few basic methods
+      #   to check for domain availability and get domain status.
+      #   Please consider to contribute implementing missing methods.
+      # 
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
+      #
+      # @since RELEASE
+      class WhoisFi < Base
+
+        property_supported :status do
+          if content_for_scanner =~ /status:\s+(.+?)\n/
+            case $1.downcase
+              when "granted"
+                :registered
+              else
+                Whois.bug!(ParserError, "Unknown status `#{$1}'.")
+            end
+          else
+            :available
+          end
+        end
+
+        property_supported :available? do
+          (status == :available)
+        end
+
+        property_supported :registered? do
+          !available?
+        end
+
+
+        property_supported :created_on do
+          if content_for_scanner =~ /created:\s+(.+)\n/
+            Time.parse($1)
+          end
+        end
+
+        property_not_supported :updated_on
+
+        property_supported :expires_on do
+          if content_for_scanner =~ /expires:\s+(.+)\n/
+            Time.parse($1)
+          end
+        end
+
+
+        property_supported :nameservers do
+          content_for_scanner.scan(/nserver:\s+(.+)\n/).flatten.map do |line|
+            Record::Nameserver.new(line.split(" ").first)
+          end
+        end
+
+      end
+
+    end
+  end
+end