+++ lib/whois/record/parser/whois.ficora.fi.rb
@@ -7,69 +7,17 @@
 #++
 
 
-require 'whois/record/parser/base'
+require 'whois/record/parser/whois.fi'
 
 
 module Whois
   class Record
     class Parser
 
-      #
-      # = whois.ficora.fi parser
-      #
       # Parser for the whois.ficora.fi server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
-      #
-      class WhoisFicoraFi < Base
-
-        property_supported :status do
-          if content_for_scanner =~ /status:\s+(.+?)\n/
-            case $1.downcase
-              when "granted" then :registered
-              else
-                Whois.bug!(ParserError, "Unknown status `#{$1}'.")
-            end
-          else
-            :available
-          end
-        end
-
-        property_supported :available? do
-          (status == :available)
-        end
-
-        property_supported :registered? do
-          !available?
-        end
-
-
-        property_supported :created_on do
-          if content_for_scanner =~ /created:\s+(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-        property_not_supported :updated_on
-
-        property_supported :expires_on do
-          if content_for_scanner =~ /expires:\s+(.*)\n/
-            Time.parse($1)
-          end
-        end
-
-
-        property_supported :nameservers do
-          content_for_scanner.scan(/nserver:\s+(.+)\n/).flatten.map do |line|
-            Record::Nameserver.new(line.split(" ").first)
-          end
-        end
-
-      end
+      # It aliases the whois.fi parser.
+      WhoisFicoraFi = WhoisFi
 
     end
   end