+++ CHANGELOG.md
@@ -3,16 +3,24 @@
 
 ## master
 
-* SERVER: .KP TLD no longer provides public WHOIS interface (GH-52).
+* SERVER: Removed .KP (GH-52), .GP (GH-44) TLD servers.
 
-* SERVER: Added the following new IDN TLDs:
+* SERVER: Updated .CV, .FI (GH-133), .GF, .MQ, .UA (GH-153), .XN-MGBAYH7GPA TLD servers.
+
+* SERVER: Added the following new IDN TLD servers:
     * .XN--80AO21A (.ҚАЗ, Kazakhstan)
+
+* SERVER: Updated ipv4 allocations.
+
+* SERVER: Added .RS, .LK TLD server.
 
 * NEW: Added full whois.dreamhost.com parser.
 
 * NEW: Added ability to check for an unavailable response with Whois::Record#response_unavailable?.
 
 * NEW: whois.nic.uk now understands reserved domains.
+
+* NEW: Added full whois.rnids.rs parser.
 
 * CHANGED: Placed "core_ext" under "whois" namespace (GH-149).